实现Spring Schedule和XxlJob定时任务来完成文章和评论计数缓存的同步

+++ src/main/java/org/wj/letsrock/domain/article/repository/ArticleRepository.java
@@ -52,5 +52,5 @@ public interface ArticleRepository extends IService<ArticleDO> {
 
     Page<ArticleDO> listHotArticles(PageParam pageParam);
 
-    void updateArticleStatisticInfo(Long articleId, Map<String, Integer> statistics);
+    void updateArticleStatisticInfo(Long articleId, Map<String, Long> statistics);
 }
